fix(#63): Service 클래스 인터페이스,impl로 분리
